scripts/requirements.txt
PyGithub
pydantic
Jinja2
python-frontmatter
